refactor(audit-dialog): extract get audit summary into custom hook
